refactor(editor): 塗り・スタンプ領域の Konva ノードを分割
EditorCanvas の責務を薄くするため、塗り領域・スタンプ領域・エフェクトプレビューをコンポーネント化し、スタンプ画像選択を lib 化して単体テストを追加する。
